Add correct projects information and enhancements to contact flow
The projects now have the correct description and images. I added a link
to the Project Card component in order to let the user visit the
project's published page.

The contact page now has a smooth animation that hides the 'repeated'
header link by making them 0 opacity and then applying a display:
'none'. Also the contact form now redirects to a new 'success' page
after submitting where the user is prompted to go back to the home page.